% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate_diffmeans.R
\name{ate_diffmeans}
\alias{ate_diffmeans}
\title{Simple Difference in Means estimator for binary treatment (for HW etc)}
\usage{
ate_diffmeans(df, outcome, treatment)
}
\arguments{
\item{df}{A dataframe}

\item{outcome}{The outcome variable (in df)}

\item{treatment}{The treatment indicator (in df)}
}
\description{
Simple Difference in Means estimator for binary treatment (for HW etc)
}
\examples{
nsw_te = ate_diffmeans(nsw_exper, re78, nsw)
}
\keyword{ATE}
\keyword{Effect}
\keyword{Treatment}
