% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\name{balance_table}
\alias{balance_table}
\title{Balance Check}
\usage{
balance_table(df, treatvar, bal_vars)
}
\arguments{
\item{df}{dataframe to check balance}

\item{treatvar}{dummy for treatment}

\item{bal_vars}{covariates to check balance in}
}
\description{
Balance Check
}
\examples{
formula_lfe(y='mpg', X = c('hp', 'drat'), D = c('wt', 'vs'))
balance_table(nsw_psid, 'nsw', c('age', 'black', 'u75'))
}
\keyword{balance}
\keyword{check}
\keyword{linear}
\keyword{model}
\keyword{regression}
