% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootUtils.R
\name{bboot}
\alias{bboot}
\title{Bayesian bootstrap for linear regression}
\usage{
bboot(object, reps = 100L, cluster = NULL, ...)
}
\arguments{
\item{object}{model object of class lm or fixest}

\item{reps}{number of replications}

\item{cluster}{clustering variable passed as bare formula or char vector}
}
\value{
bboot object
}
\description{
Bayesian bootstrap for linear regression
}
\examples{
lm(mpg ~ wt + hp, mtcars) |>
  bboot(reps = 1e3) |>
  summary()
}
