% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatter.R
\name{binscatter}
\alias{binscatter}
\title{R translation of Michael Stepner's binscatter for Stata
forked from timlrx -
https://github.com/timlrx/binscatter/blob/master/binscatter.R
and added some options, made it return object so labels can be
customised later}
\usage{
binscatter(formula, key_var, data, plotraw = TRUE, bins = 20,
  partial = FALSE)
}
\arguments{
\item{formula}{, key_var, dataframe, number of bins, option for partials}
}
\description{
R translation of Michael Stepner's binscatter for Stata
forked from timlrx -
https://github.com/timlrx/binscatter/blob/master/binscatter.R
and added some options, made it return object so labels can be
customised later
}
\examples{
data("Guns", package = "AER")
binscatter(formula="violent ~ prisoners", key_var = "prisoners",
          data=Guns, bins=10, partial=FALSE)
}
\keyword{bins}
\keyword{cef}
\keyword{scatterplot}
