% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlUtils.R
\name{calibPlot}
\alias{calibPlot}
\title{plot calibration}
\usage{
calibPlot(y, yhat, nbins = 10)
}
\arguments{
\item{y}{true value (vector)}

\item{yhat}{predicted value (vector)}

\item{nbins}{number of bins (defaults to 10)}
}
\value{
ggplot of calibration
}
\description{
plot calibration
}
