% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlUtils.R
\name{classifEval}
\alias{classifEval}
\title{Classification Evaluation}
\usage{
classifEval(actual, predicted)
}
\arguments{
\item{actual}{vector of real labels}

\item{predicted}{vector of predicted labels}
}
\value{
list of classification accuracy metrics
}
\description{
Classification Evaluation
}
