% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliners.R
\name{clip}
\alias{clip}
\title{clip vector to specified range (clone of np.clip)}
\usage{
clip(x, l, u)
}
\arguments{
\item{x}{vector}

\item{l}{lower bound}

\item{u}{upper bound}
}
\value{
clipped vector
}
\description{
clip vector to specified range (clone of np.clip)
}
