% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_all_to_csv.R
\name{convert_all_to_csv}
\alias{convert_all_to_csv}
\title{Writes all dataframes to csv}
\usage{
convert_all_to_csv(dataframes, path)
}
\arguments{
\item{dataframes}{= list of dataframes,}

\item{path}{= path of file/s}
}
\description{
Writes all dataframes to csv
}
\examples{
convert_all_to_csv(dataframes, path)
}
\keyword{bulk}
\keyword{save}
\keyword{write}
