% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlUtils.R
\name{createFolds}
\alias{createFolds}
\title{create folds for cross-fitting}
\usage{
createFolds(df, nf = 10)
}
\arguments{
\item{df}{dataframe - used to compute nrows}

\item{nf}{Number of folds}
}
\value{
list with indices for each fold in first element, and fold assignment vector in second element
}
\description{
create folds for cross-fitting
}
