% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliners.R
\name{dataPkg}
\alias{dataPkg}
\title{list datasets in package}
\usage{
dataPkg(pkg)
}
\arguments{
\item{pkg}{package name in string}
}
\value{
table with names of data objects
}
\description{
list datasets in package
}
