% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Detatches all but base R packages}
\usage{
detachAllPackages()
}
\description{
Detatches all but base R packages
}
\examples{
detachAllPackages()
}
\keyword{libraries}
\keyword{reset}
