% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalFns.R
\name{DMLReg}
\alias{DMLReg}
\title{Simple Double-ML implementation with cross-fitting}
\usage{
DMLReg(x, d, y, dreg, yreg, nfold = 5)
}
\arguments{
\item{x}{Matrix of covariates}

\item{d}{treatment variable}

\item{y}{The dependent variable}

\item{dreg}{function that wraps learner - e.g. glmnet - for pscore}

\item{yreg}{function that wraps learner - e.g. glmnet - for outcome regression}

\item{D}{vector of factor variables to be partialed out}

\item{nfolds}{number of folds for cross fitting}
}
\description{
Simple Double-ML implementation with cross-fitting
}
\examples{
\dontrun{
dreg <- function(x,d){ cv.glmnet(x, d, alpha = 1) }
yreg <- function(x,y){ cv.glmnet(x, y, alpha = 1) }
res <- DMLReg( x=x, d=d, y=y, dreg=dreg, yreg=yreg, nfold=5)
}
}
\keyword{causal}
\keyword{double}
\keyword{inferenc}
\keyword{ml,}
