% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_missing_vars.R
\name{drop_missing_vars}
\alias{drop_missing_vars}
\title{Drop variables with all NAs in dataframe}
\usage{
drop_missing_vars(df)
}
\arguments{
\item{dataframe}{A dataframe}
}
\value{
dataframe with empty variables dropped
}
\description{
Drop variables with all NAs in dataframe
}
\examples{
clean_dataframe = drop_missing_vars(dataframe_with_empty_vars)
}
\keyword{dataframe}
\keyword{variable}
