% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regUtils.R
\name{fitGet}
\alias{fitGet}
\title{return cross-fit predictions from glmnet object}
\usage{
fitGet(m)
}
\arguments{
\item{m}{glmnet model object fit with keep = T}
}
\value{
vector of cross-fit predictions
}
\description{
return cross-fit predictions from glmnet object
}
