% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaMakers.R
\name{fmla_kth}
\alias{fmla_kth}
\title{construct formula with outcome and polynomial terms}
\usage{
fmla_kth(y, b, c, k = 2)
}
\arguments{
\item{y}{outcome}

\item{b}{binary variables}

\item{c}{continuous vars (for polynomials)}

\item{k}{polynomial order}
}
\value{
formula
}
\description{
construct formula with outcome and polynomial terms
}
