% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_lfe.R
\name{formula_lfe}
\alias{formula_lfe}
\title{Stitches together formula for use in felm}
\usage{
formula_lfe(y, X, w = NULL, D = NULL, Z = NULL, C = NULL)
}
\arguments{
\item{y}{The dependent variable}

\item{X}{vector of Xs}

\item{w}{endogenous variable (only included if first stage is specified)}

\item{D}{vector of factor variables to be partialed out}

\item{Z}{vector of instruments}

\item{C}{vector of variables cluster standard errors}
}
\description{
Stitches together formula for use in felm
}
\examples{
formula_lfe(y='mpg', X = c('hp', 'drat'), D = c('wt', 'vs'))
formula_lfe(y='mpg', X = c('hp', 'drat'), w = 'gear', Z = c('cyl', 'carb'), D = c('wt', 'vs'), C = c('cyl', 'wt'))
}
\keyword{linear}
\keyword{model}
\keyword{name}
\keyword{regression}
\keyword{variable}
