% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_stitcher.R
\name{formula_stitcher}
\alias{formula_stitcher}
\title{Stitches together formula for use in lm/glm}
\usage{
formula_stitcher(depvar, rhs_vars, factors = NULL)
}
\arguments{
\item{depvar}{The dependent variable}

\item{rhs_vars}{List of independent variables (continuous/dummy variables)}

\item{factors}{List of factor variables (NULL by default)}
}
\description{
Stitches together formula for use in lm/glm
}
\examples{
fml1 = formula_stitcher('wage',c('age','experience','married'),c('ethnicity','sector'))
lm1 <- lm(fml1,data=CPS1985)
}
\keyword{dataframe}
\keyword{name}
\keyword{variable}
