% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Summary table with percentages for categorical variables}
\usage{
freq_table(df, var)
}
\arguments{
\item{df}{A dataframe , var A column in said dataframe}
}
\description{
Summary table with percentages for categorical variables
}
\examples{
freq_table(mtcars, 'as.factor(cyl)')
}
\keyword{categorical}
\keyword{dataframe}
\keyword{name}
\keyword{variable}
