% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreq.R
\name{fromLimportF}
\alias{fromLimportF}
\title{Loads specific function or list of functions f from library l}
\usage{
fromLimportF(l, f)
}
\arguments{
\item{l}{library name (needs to be installed)}

\item{f}{function name}
}
\description{
Loads specific function or list of functions f from library l
}
