% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{fuzzy matches 2 character vectors, and returns most similar pairs}
\usage{
fuzzy_match(a, b)
}
\arguments{
\item{character}{vectors a and b}
}
\description{
fuzzy matches 2 character vectors, and returns most similar pairs
}
\examples{
fuzzy_match(keys_A, keys_B)
}
\keyword{fuzzy}
\keyword{match}
\keyword{string}
