% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{get_and_unzip}
\alias{get_and_unzip}
\title{downloads and unzips file if it is not found in current wdir}
\usage{
get_and_unzip(url, filename)
}
\arguments{
\item{url, }{filename}
}
\description{
downloads and unzips file if it is not found in current wdir
}
\examples{
\dontrun{
get_and_unzip('url','qob.txt')
}
}
\keyword{download}
