% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructUtils.R
\name{getNthElement}
\alias{getNthElement}
\title{Get n'th element in list of lists}
\usage{
getNthElement(list.of.lists, nth.element)
}
\arguments{
\item{list.of.lists}{list of lists}

\item{nth.element}{index of list to return}
}
\value{
list
}
\description{
Get n'th element in list of lists
}
