% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regUtils.R
\name{interSparseM}
\alias{interSparseM}
\title{prepare sparse X matrix with interactions}
\usage{
interSparseM(data, k, corr_cut = 0.9)
}
\arguments{
\item{data}{data table / dataframe}

\item{k}{order of interactions: defaults to pairwise interactions}
}
\value{
matrix with interactions
}
\description{
Prepare sparse matrix with  interactions.
}
