% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgps.R
\name{KS2007_dgp}
\alias{KS2007_dgp}
\title{Kang Schaefer (2007) DGP for 'mild' confounding}
\usage{
KS2007_dgp(n = 1000, τ = 10)
}
\arguments{
\item{n}{nobs}

\item{τ}{true effect (10 by default)}
}
\description{
Kang Schaefer (2007) DGP for 'mild' confounding
}
