% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_extractor.R
\name{label_extractor}
\alias{label_extractor}
\title{extracts variable labels and returns a dataframe with varname-varlabel for DTA files}
\usage{
label_extractor(df, colnames = c("names", "var.labels"))
}
\arguments{
\item{dataframe}{(read from stata)}
}
\description{
extracts variable labels and returns a dataframe with varname-varlabel for DTA files
}
\examples{
label_extractor(df)
}
\keyword{data}
\keyword{labels}
\keyword{stata}
