% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lal_plot_theme.R
\name{lal_plot_theme}
\alias{lal_plot_theme}
\title{Theme for nice looking plots}
\usage{
lal_plot_theme(fontfam = "Roboto Condensed", textangle = 0, ...)
}
\arguments{
\item{ggplot}{object}
}
\description{
Theme for nice looking plots
}
\examples{
plot_ggp_object + lal_plot_theme()
}
\keyword{ggplot}
\keyword{plots}
