% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lal_plot_theme_d.R
\name{lal_plot_theme_d}
\alias{lal_plot_theme_d}
\title{Fork of hrbrthemes::theme_ft_rc() with altered legend options and dots}
\usage{
lal_plot_theme_d(
  fontfam = "IBM Plex Sans Condensed",
  fsize = 15,
  textangle = 0,
  ...
)
}
\arguments{
\item{ggplot}{object}
}
\description{
Fork of hrbrthemes::theme_ft_rc() with altered legend options and dots
}
\examples{
plot_ggp_object + lal_plot_theme_d()
}
\keyword{ggplot}
\keyword{plots}
