% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreq.R
\name{libreq}
\alias{libreq}
\title{Loads a list of libraries, installs when necessary - Take 2}
\usage{
libreq(..., repo = "https://cloud.r-project.org/")
}
\arguments{
\item{repo}{string containing repo to be passed to install.packages}
}
\description{
Loads a list of libraries, installs when necessary - Take 2
}
\examples{
libreq(tidyverse, lfe, estimatr)
}
\keyword{install}
\keyword{libraries}
