% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructUtils.R
\name{listOmitIf}
\alias{listOmitIf}
\title{Omit element if it contains "contains"}
\usage{
listOmitIf(lst, contains)
}
\arguments{
\item{y}{list of elements}
}
\value{
y with "contains" omitted
}
\description{
Omit element if it contains "contains"
}
