% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install.R
\name{load_or_install}
\alias{load_or_install}
\title{Loads a list of libraries, installs when necessary}
\usage{
load_or_install(wants, repo = "https://cloud.r-project.org/")
}
\arguments{
\item{wants}{A list of R modules available on CRAN}

\item{repo}{string containing repo to be passed to install.packages}
}
\description{
Loads a list of libraries, installs when necessary
}
\examples{
load_or_install(c('tidyverse','Hmisc','glmnet'))
}
\keyword{install}
\keyword{libraries}
