% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_varnames.R
\name{lower_varnames}
\alias{lower_varnames}
\title{Lower variable names in a dataframe}
\usage{
lower_varnames(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Lower variable names in a dataframe
}
\examples{
lower_varnames(dataframe_with_bad_varnames)
}
\keyword{dataframe}
\keyword{name}
\keyword{variable}
