% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliners.R
\name{maskMatrix}
\alias{maskMatrix}
\title{mask a matrix using another matrix}
\usage{
maskMatrix(A, M)
}
\arguments{
\item{A}{matrix}

\item{M}{Mask matrix (must be same dimensions with a rectangular subset of TRUEs)}
}
\value{
masked matrix
}
\description{
mask a matrix using another matrix
}
