% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{mcReplicate}
\alias{mcReplicate}
\title{Multi-core replicate.
Use multiple cores for repeated evaluation of an expression.}
\usage{
mcReplicate(n, expr, mc.cores = 4, refresh = 0.1)
}
\arguments{
\item{n}{integer; the number of replications.}

\item{expr}{the expression (a language object, usually a call) to evaluate n times.}

\item{mc.cores}{number of cores to use.}

\item{refresh}{status update refresh interval.}
}
\value{
A vector, matrix, or list of length `n`.
}
\description{
Multi-core replicate.
Use multiple cores for repeated evaluation of an expression.
}
\examples{
one_sim = \(n = 100, control_prob = 0.1, rel_effect = 0.01) {
  treat_prob = control_prob + (control_prob * rel_effect)
  cy = rbinom(n = n, size = 1, prob = control_prob)
  ty = rbinom(n = n, size = 1, prob = treat_prob)
  mean(ty) - mean(cy)
  }
  mcReplicate(10, one_sim(), mc.cores = 2)

}
