% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{mcSapply}
\alias{mcSapply}
\title{Multi-core sapply
Use multiple cores for simplifying apply.}
\usage{
mcSapply(X, FUN, mc.cores = 6, ..., simplify = TRUE)
}
\arguments{
\item{X}{input vector to iterate over}

\item{FUN}{function}

\item{mc.cores}{number of cores to use.}
}
\value{
A vector, matrix, or list of length `n`.
}
\description{
Multi-core sapply
Use multiple cores for simplifying apply.
}
