% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Stitches together multiple ggplot objects for export-ready graphs}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{ggplot}{objects}
}
\description{
Stitches together multiple ggplot objects for export-ready graphs
}
\examples{
plot1 = ggplot(data,aes(xvar1,yvar)) + geom_point()
plot2 = ggplot(data,aes(xvar2,yvar)) + geom_point()
multiplot(plot1,plot2,cols=2)
}
\keyword{graphs}
\keyword{plots}
