% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathsUtils.R
\name{newtonRoot}
\alias{newtonRoot}
\title{solve for roots using the NR update rule x1 = x0 − f(x0)/f'(x0)}
\usage{
newtonRoot(f, df, x0, tol = 0.00001, maxit = 20)
}
\description{
solve for roots using the NR update rule x1 = x0 − f(x0)/f'(x0)
}
