% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss_nun.R
\name{nmiss_nun}
\alias{nmiss_nun}
\title{Table of missing value counts and number of unique values}
\usage{
nmiss_nun(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Table of missing value counts and number of unique values
}
\examples{
nmiss_nun(mtcars)
}
\keyword{dataframe}
\keyword{missing}
\keyword{name}
\keyword{variable}
