% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_diff.R
\name{norm_diff}
\alias{norm_diff}
\title{Normalised Difference (Imbens, Rubin 2015) for observational studies}
\usage{
norm_diff(df, treat, value)
}
\arguments{
\item{df}{dataframe to check balance}

\item{treat}{dummy for treatment}

\item{value}{covariate to calculate normalised difference statistic}
}
\description{
Normalised Difference (Imbens, Rubin 2015) for observational studies
}
\examples{
norm_diff(mtcars, vs, gear)
}
\keyword{balance}
\keyword{check}
\keyword{linear}
\keyword{model}
\keyword{regression}
