% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliners.R
\name{nunique}
\alias{nunique}
\title{number of unique values in vector}
\usage{
nunique(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar with number of unique values
}
\description{
number of unique values in vector
}
