% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalFns.R
\name{OLSw}
\alias{OLSw}
\title{Compute OLS with arbitrary vector of weights (possibly negative)}
\usage{
OLSw(y, X, w)
}
\arguments{
\item{y}{Response vector}

\item{X}{A numeric data matrix}

\item{w}{vector of weights (same length as y)}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Compute OLS with arbitrary vector of weights (possibly negative)
}
