% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{overlap_histogram}
\alias{overlap_histogram}
\title{Histogram of pscore distributions in treatment and control groups}
\usage{
overlap_histogram(w, ps)
}
\arguments{
\item{w}{treatment vector}

\item{ps}{propensity score vector}
}
\value{
figure with overlap
}
\description{
Histogram of pscore distributions in treatment and control groups
}
\examples{
data(lalonde.exp)
pscore = glm(treat ~ .-re78, lalonde.exp, family = binomial)$fitted.values
overlap_histogram(lalonde.exp$treat, pscore) + lal_plot_theme()
}
