% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_treat_plot.R
\name{panel_treat_plot}
\alias{panel_treat_plot}
\title{plot binary treatment status for time series for panel dataset}
\usage{
panel_treat_plot(data, time, id, status)
}
\arguments{
\item{data}{dataframe}

\item{time}{bare name of time (not in quotes)}

\item{id}{bare name of unit}

\item{status}{bare name of treatment}
}
\description{
plot binary treatment status for time series for panel dataset
}
\examples{
\dontrun{
panel_treat_plot(california_prop99, Year, State, treated)
}
}
\keyword{dataframe}
\keyword{diff-in-diff}
\keyword{panel}
