% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regUtils.R
\name{panelMatrices}
\alias{panelMatrices}
\title{Convert a long panel to wide matrix (cleaner data.table version)}
\usage{
panelMatrices(dt, unit_id, time_id, treat, outcome)
}
\arguments{
\item{dt}{data.table in long panel format}

\item{unit_id}{unit id name}

\item{time_id}{time id name}

\item{treat}{treatement name}

\item{outcome}{outcome name}
}
\value{
list with treatment matrix W, outcome matrix Y, N0 (number of control units), and T0 (number of untreated periods)
}
\description{
Convert a long panel to wide matrix (cleaner data.table version)
}
