% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regUtils.R
\name{prepBMatrix}
\alias{prepBMatrix}
\title{prepare X matrix with flexible interactions and polynomials}
\usage{
prepBMatrix(dat, continuouses, dummies = NULL, corr_cut = 0.98, k = 2, m = 2)
}
\arguments{
\item{dat}{data table / dataframe}

\item{continuouses}{Names of continuous vars to modify fn form}

\item{dummies}{Names of dummy vars}

\item{corr_cut}{cutoff for correlation threshold to drop one of the vars (default = 0.98)}

\item{k}{order of interactions: defaults to pairwise interactions}

\item{m}{order of functions: defaults to quadratic functions}
}
\value{
data.frame with base terms and interactions + basis
}
\description{
Prepare matrix with polynomial basis expansion and/or interactions
}
\examples{
data(lalonde.psid)
xn = setdiff(colnames(lalonde.psid) , c("treat", "re78"))
co = c("age", "education", "re74", "re75")
bi = c("black", "hispanic", "married", "nodegree", "u74", "u75")
Xx = prepBMatrix(lalonde.psid[, xn], co, bi)
colnames(lalonde.psid[, xn])  |> length()
colnames(Xx)                  |> length()
}
