% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{pums}
\alias{pums}
\title{Public use wage, education data from 1980 census.
Used in Mostly Harmless for Mincer Regression.}
\format{
A data frame with 329,509 rows and 6 variables
}
\usage{
data(pums)
}
\description{
\itemize{
  \item \code{lwklywge} - Log Wage
  \item \code{educ} - number of years of schooling.
  \item \code{yob} - year of birth
  \item \code{qob} - quarter of birth
  \item \code{pob} - place of birth (state)
  \item \code{age} - age in 1980
}
}
