% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_export.R
\name{rdr_export}
\alias{rdr_export}
\title{Summary table for Rdrobust estimates to export using kableExtra}
\usage{
rdr_export(rdr_out, prec = 3)
}
\arguments{
\item{rdr_out}{Rdrobust object}

\item{prec}{decimal places}
}
\description{
Summary table for Rdrobust estimates to export using kableExtra
}
\examples{
rdrobust(df$vote, df$margin, all = T) \%>\% rdr_export -> rdre
rdre \%>\% kable(, format = "latex", booktabs = T) \%>\% collapse_rows(columns = 1, latex_hline = "major", valign = "middle") \%>\% cat(., file = "table.tex")
}
\keyword{export}
\keyword{regression-discontinuity}
