% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_files.R
\name{read_all_files}
\alias{read_all_files}
\title{Reads in all datasets in given location}
\usage{
read_all_files(extension, location)
}
\arguments{
\item{path}{of file/s}
}
\description{
Reads in all datasets in given location
}
\examples{
read_all_files(extension='dta',path='~/data/')
}
\keyword{read}
\keyword{sas}
\keyword{spss}
\keyword{stata}
