% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalFns.R
\name{reg_adjust}
\alias{reg_adjust}
\title{Regression Adjustment estimation of ATE or ATT}
\usage{
reg_adjust(df, w, y, xs, estimand = "ATT")
}
\arguments{
\item{df}{dataframe}

\item{w}{treatment name}

\item{y}{outcome name}

\item{xs}{vector of covariate names}

\item{estimand}{: runs Lin regression when "ATE" and Oaxaca-Blinder-Kitagawa when "ATT"}
}
\value{
fixest model object with robust standard errors (can be summarised again with different clusters)
}
\description{
Regression Adjustment estimation of ATE or ATT
}
\examples{
data(lalonde.psid); data(lalonde.exp);
y = "re78"; w = "treat"; xs = setdiff(colnames(lalonde.psid), c(y, w))
cat("ATE in experimental sample \n")
reg_adjust(lalonde.exp, 'treat', 're78', xs, "ATE")
cat(" --------------------------- \n")
cat("ATT in obs sample \n")
reg_adjust(lalonde.psid, 'treat', 're78', xs, "ATT")
cat(" --------------------------- \n")
}
