% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualise.R
\name{residualise}
\alias{residualise}
\title{Partial out controls and fixed effects and return residualised outcome and treatment}
\usage{
residualise(y, a, x = "1", d = "0", df)
}
\arguments{
\item{y}{outcome}

\item{a}{primary rhs variable}

\item{x}{controls}

\item{d}{fixed effects}

\item{df}{dataframe}
}
\description{
Partial out controls and fixed effects and return residualised outcome and treatment
}
\examples{
\dontrun{
residualise('mpg', 'wt', 'cyl', mtcars)
}
}
\keyword{Frisch-Waugh-Lovell}
\keyword{out}
\keyword{partial}
