% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se.R
\name{robust_se}
\alias{robust_se}
\title{Returns summary for lm object with robust standard errors}
\usage{
robust_se(model)
}
\arguments{
\item{lm}{object}
}
\description{
Returns summary for lm object with robust standard errors
}
\examples{
robust_se(lm(mpg ~ wt, mtcars))
}
\keyword{hc0}
\keyword{heteroskedasticity}
\keyword{robust}
