% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{robust_summary}
\alias{robust_summary}
\title{Returns summary for lm object with robust standard errors}
\usage{
robust_summary(model)
}
\arguments{
\item{lm}{object}
}
\description{
Returns summary for lm object with robust standard errors
}
\examples{
robust_summary(lm(y~x,data=df))
}
\keyword{hc0}
\keyword{heteroskedasticity}
\keyword{robust}
