% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regUtils.R
\name{robustify}
\alias{robustify}
\title{Replaces the standard errors (t and p vals)in FELM model object with robust SE}
\usage{
robustify(model)
}
\arguments{
\item{felm}{object}
}
\description{
Replaces the standard errors (t and p vals)in FELM model object with robust SE
}
\examples{
\dontrun{
robustify(felm(y ~ x, data = df))
}
}
\keyword{hc0}
\keyword{heteroskedasticity}
\keyword{output}
\keyword{robust}
