% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_maker.R
\name{se_maker}
\alias{se_maker}
\title{Returns summary for lm object with robust standard errors}
\usage{
se_maker(vcovmat)
}
\arguments{
\item{vcov}{matrix from sandwich library}
}
\description{
Returns summary for lm object with robust standard errors
}
\examples{
se_maker(vcovCL(m, cluster = df$clustervar))
}
\keyword{cluster}
\keyword{hc0}
\keyword{heteroskedasticity}
\keyword{robust}
