% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclassify.R
\name{subclass_est}
\alias{subclass_est}
\title{Compute subclassification ATT and ATE}
\usage{
subclass_est(df, x, y = "re78", w = "treat", debug = F)
}
\arguments{
\item{df}{data.table}

\item{x}{character vector of discrete covariates}

\item{y}{outcome}

\item{w}{treatment}

\item{debug}{boolean to return strata-specific estimates}
}
\description{
Compute subclassification ATT and ATE
}
