% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalFns.R
\name{subclassify}
\alias{subclassify}
\title{Compute subclassification ATT and ATE}
\usage{
subclassify(df, x, y = "re78", w = "treat", debug = F)
}
\arguments{
\item{df}{data.table}

\item{x}{character vector of discrete covariates}

\item{y}{outcome}

\item{w}{treatment}

\item{debug}{boolean to return strata-specific estimates}
}
\description{
Compute subclassification ATT and ATE
}
\examples{
\dontrun{
subclassify(mtcars, x = "cyl", y = "mpg", w = "foreign")
}
}
