% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumStats.R
\name{summar}
\alias{summar}
\title{Summary table}
\usage{
summar(df, prec = 2)
}
\arguments{
\item{df}{A dataframe , var A column in said dataframe}
}
\value{
table with summary statistics
}
\description{
Summary table
}
\examples{
summar(mtcars)
}
