% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootUtils.R
\name{summary.bboot}
\alias{summary.bboot}
\title{summary method for bayesian bootstrap}
\usage{
\method{summary}{bboot}(object, level = 0.95, ...)
}
\arguments{
\item{object}{bboot object}

\item{level}{test level (0.95 by default)}
}
\description{
summary method for bayesian bootstrap
}
