% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneliners.R
\name{timeStamp}
\alias{timeStamp}
\title{Timestamp with no spaces for filenames}
\usage{
timeStamp(ff = "\%d_\%m_\%Y__\%H\%M")
}
\value{
character with time stamp in format ff
}
\description{
Timestamp with no spaces for filenames
}
\examples{
timeStamp()
}
