% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{transfer_data}
\alias{transfer_data}
\title{Populates a stat-transfer script to convert files to CSV}
\usage{
transfer_data(
  path,
  stat.transfer.path = "\\"C:\\\\Program Files\\\\StatTransfer12-64\\\\st.exe\\"",
  out.ext = "csv",
  outpath
)
}
\arguments{
\item{path}{of file/s}
}
\description{
Populates a stat-transfer script to convert files to CSV
}
\examples{
transfer_data(path='C:/data/data1.sas7bdat',outpath='C:/working/')
}
\keyword{convert}
\keyword{reader}
\keyword{stat_transfer}
