% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplotter.R
\name{tsplotter}
\alias{tsplotter}
\title{Canned subroutine to plot time series for several variables}
\usage{
tsplotter(df, timevar = "year")
}
\arguments{
\item{df}{A dataframe}

\item{timevar}{a string with a colname for the time variable}
}
\description{
Canned subroutine to plot time series for several variables
}
\examples{
overall_ineq = tsplotter(overall) + labs(title='Overall Inequality')
}
\keyword{categorical}
\keyword{dataframe}
\keyword{name}
\keyword{variable}
