% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlUtils.R
\name{ttsplit}
\alias{ttsplit}
\title{test-train split}
\usage{
ttsplit(df, k = 0.75)
}
\arguments{
\item{df}{[data.frame, matrix] data table (n x p)}

\item{k}{share of data obs in train}
}
\value{
list with test and train
}
\description{
test-train split
}
