% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{withFrame}
\alias{withFrame}
\title{ggplot with frame}
\usage{
withFrame(g, aes, d, ...)
}
\arguments{
\item{g}{geom}

\item{aes}{ggplot2 aesthetic}

\item{d}{dataframe}
}
\value{
populated geom
}
\description{
ggplot with frame
}
