% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\name{balance_table}
\alias{balance_table}
\title{Balance Check for observational studies; includes SD and Normalised Difference}
\usage{
balance_table(df, treatvar, bal_vars)
}
\arguments{
\item{df}{dataframe to check balance}

\item{treatvar}{dummy for treatment}

\item{bal_vars}{covariates to check balance in}
}
\description{
Balance Check for observational studies; includes SD and Normalised Difference
}
\examples{
balance_table(mtcars, treatvar = 'vs', bal_vars = c('mpg', 'cyl', 'disp'))
}
\keyword{balance}
\keyword{check}
\keyword{linear}
\keyword{model}
\keyword{regression}
