% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatter.R
\name{binscatter}
\alias{binscatter}
\title{R translation of Michael Stepner's binscatter for Stata
forked from timlrx -
https://github.com/timlrx/binscatter/blob/master/binscatter.R
and added some options, made it return object so labels can be
customised later}
\usage{
binscatter(formula, key_var, data, plotraw = TRUE, bins = 20,
  partial = T)
}
\arguments{
\item{formula}{FELM formula}

\item{key_var}{W variable for CEF}

\item{data}{dataframe}

\item{plotraw}{T if underlying scatterplot should be plotted}

\item{bins}{= number of bins}

\item{partial}{T if Y and W should be residualised on other Xs first}
}
\description{
R translation of Michael Stepner's binscatter for Stata
forked from timlrx -
https://github.com/timlrx/binscatter/blob/master/binscatter.R
and added some options, made it return object so labels can be
customised later
}
\examples{
data("Guns", package = "AER")
binscatter(formula="violent ~ prisoners", key_var = "prisoners",
          data=Guns, bins=10, partial=F)
}
\keyword{bins}
\keyword{cef}
\keyword{scatterplot}
