% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_vlabs.R
\name{dta_vlabs}
\alias{dta_vlabs}
\title{Small summary table with stata variable labels from DTA file}
\usage{
dta_vlabs(dta)
}
\arguments{
\item{df}{dataframe to check balance}
}
\description{
Small summary table with stata variable labels from DTA file
}
\examples{
dta_vlabs(df)
}
\keyword{dataframe}
\keyword{labels}
\keyword{stata}
\keyword{summary}
