% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_stitcher.R
\name{formula_stitcher}
\alias{formula_stitcher}
\title{Stitches together formula for use in lm/glm}
\usage{
formula_stitcher(Y, X, factors = NULL)
}
\arguments{
\item{Y}{The dependent variable}

\item{X}{List of independent variables (continuous/dummy variables)}

\item{factors}{List of factor variables (NULL by default)}
}
\description{
Stitches together formula for use in lm/glm
}
\examples{
fml1 = formula_stitcher('wage',c('age','experience','married'),c('ethnicity','sector'))
}
\keyword{dataframe}
\keyword{name}
\keyword{variable}
