% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_unzip.R
\name{get_and_unzip}
\alias{get_and_unzip}
\title{downloads and unzips file if it is not found in current wdir}
\usage{
get_and_unzip(url, filename)
}
\arguments{
\item{url, }{filename}
}
\description{
downloads and unzips file if it is not found in current wdir
}
\examples{
get_and_unzip('link_to_angrist_site','qob.txt')
}
\keyword{download}
