% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lal_map_theme.R
\name{lal_map_theme}
\alias{lal_map_theme}
\title{Theme for maps}
\usage{
lal_map_theme(fontfam = "Roboto Condensed", ...)
}
\arguments{
\item{ggplot}{object}
}
\description{
Theme for maps
}
\examples{
map_ggp_object + lal_map_theme()
}
\keyword{ggplot}
\keyword{maps}
\keyword{spatial}
