% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{lal_plot_theme}
\alias{lal_plot_theme}
\title{theme for nice looking plots with sensible defaults (no legend title, legend at the bottom)}
\usage{
lal_plot_theme(
  fontfam = "IBM Plex Sans Condensed",
  fsize = 14,
  textangle = 0,
  ...
)
}
\arguments{
\item{ggplot}{object}
}
\description{
theme for nice looking plots with sensible defaults (no legend title, legend at the bottom)
}
\examples{
\dontrun{
plot_ggp_object + lal_plot_theme()
}
}
\keyword{ggplot}
\keyword{plots}
