% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustify.R
\name{robustify}
\alias{robustify}
\title{Replaces the standard errors (t and p vals)in FELM model object with robust
so that passing it on to stargazer spits out correct SEs
Only necessary for non-clustered SE; FELM automatically calculates
clustered standard errors if Y ~ X | FE | IV | Cluster specified}
\usage{
robustify(model)
}
\arguments{
\item{felm}{object}
}
\description{
Replaces the standard errors (t and p vals)in FELM model object with robust
so that passing it on to stargazer spits out correct SEs
Only necessary for non-clustered SE; FELM automatically calculates
clustered standard errors if Y ~ X | FE | IV | Cluster specified
}
\examples{
robustify(felm(y~x,data=df))
}
\keyword{hc0}
\keyword{heteroskedasticity}
\keyword{output}
\keyword{robust}
