% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalFns.R
\name{subclassify}
\alias{subclassify}
\title{Compute subclassification ATT and ATE}
\usage{
subclassify(df, x, y = "re78", w = "treat", debug = F)
}
\arguments{
\item{df}{data.table}

\item{x}{character vector of discrete covariates}

\item{y}{outcome}

\item{w}{treatment}

\item{debug}{boolean to return strata-specific estimates}
}
\value{
list with ATE and ATT estimate in est and strata level averages in table
}
\description{
Compute subclassification ATT and ATE
}
\examples{
data(lalonde.psid); setDT(lalonde.psid)
subclassify(lalonde.psid,
  x = c("u74", "u75"),
  y = "re78", w = "treat"
)
subclassify(lalonde.psid,
  x = c("married", "black", "hispanic", "u74", "u75"),
  y = "re78", w = "treat"
)

}
